fix: Refactor code to use double instead of float for decimal values

+++ Data/JsonTransformers/QuestionData.cs
@@ -1,7 +1,9 @@
-namespace JavaHateBE.Data.JsonTransformers {
-    public struct QuestionData {
+namespace JavaHateBE.Data.JsonTransformers
+{
+    public struct QuestionData
+    {
         public string Text { get; set; }
-        public float Answer { get; set; }
+        public double Answer { get; set; }
         public byte Difficulty { get; set; }
     }
 }
